+++ packages/ricos-common/web/src/themeStrategy/generators/customStyles.ts
@@ -1,4 +1,9 @@
-import type { CustomTextualStyle, RicosSettingsStyles, RicosCustomStyles } from 'ricos-types';
+import type {
+  CustomTextualStyle,
+  RicosSettingsStyles,
+  RicosCustomStyles,
+  RicosOneAppStyles,
+} from 'ricos-types';
 import type { CssVarsObject } from '../themeTypes';
 import { merge } from 'lodash';
 
@@ -17,7 +22,7 @@ const lineHeightFix = ({
 });
 
 type ToVars = (
-  customStyles: RicosCustomStyles | RicosSettingsStyles,
+  customStyles: RicosCustomStyles | RicosSettingsStyles | RicosOneAppStyles,
   prefix: string,
   withLineHeightFix?: boolean
 ) => CssVarsObject;
@@ -39,10 +44,20 @@ const toVars: ToVars = (customStyles, prefix, fix = false) =>
 type CreateCustomStyles = (param: {
   customStyles?: RicosCustomStyles;
   settingsStyles?: RicosSettingsStyles;
+  oneAppStyles?: RicosOneAppStyles;
 }) => CssVarsObject;
 
-const createCustomStyles: CreateCustomStyles = ({ customStyles = {}, settingsStyles = {} }) => {
-  return merge({}, toVars(customStyles, 'custom', true), toVars(settingsStyles, 'settings'));
+const createCustomStyles: CreateCustomStyles = ({
+  customStyles = {},
+  settingsStyles = {},
+  oneAppStyles = {},
+}) => {
+  return merge(
+    {},
+    toVars(customStyles, 'custom', true),
+    toVars(settingsStyles, 'settings'),
+    toVars(oneAppStyles, 'oneapp')
+  );
 };
 
 export default createCustomStyles;